+++ test/blackbox-tests/test-cases/extra-objects/extra-objects-error-on-duplicate.t
@@ -6,13 +6,13 @@ Test that duplicate foreign objects results in an error
   $ cat >dune <<EOF
   > (executable
   >  (name foo)
-  >  (foreign_objects foo foo))
+  >  (extra_objects foo foo))
   > EOF
 
   $ dune build
-  File "dune", line 3, characters 18-21:
-  3 |  (foreign_objects foo foo))
-                        ^^^
+  File "dune", line 3, characters 16-19:
+  3 |  (extra_objects foo foo))
+                      ^^^
   Error: Duplicate object name: foo. Already appears at:
   - dune:3
   [1]