+++ CHANGES.md
@@ -50,7 +50,7 @@
   directories to the include paths sourced from a file. (#6058, fixes #3993,
   @gridbugs)
 
-- Support `(foreign_objects ...)` field in `(executable ...)` and `(library
+- Support `(extra_objects ...)` field in `(executable ...)` and `(library
   ...)` stanzas (#6084, fixes #4129, @gridbugs)
 
 - Fix compilation of Dune under esy on Windows (#6109, fixes #6098, @nojb)